Move CustomHashCollections to model module
